feat: generic NavBar component
* Builds on navigation-menu from shadcn to expose a generic menu'ed
  NavBar
* Allows users to override the default menu items
