+++ apps/frontend/components/widgets/index.ts
@@ -18,6 +18,10 @@ import {
 } from "@/components/widgets/feedback-farm";
 import { Markdown, MarkdownMeta } from "@/components/widgets/markdown";
 import {
+  MarketingNavbar,
+  MarketingNavbarMeta,
+} from "@/components/widgets/marketing-navbar";
+import {
   MonacoEditor,
   MonacoEditorMeta,
 } from "@/components/widgets/monaco-editor";
@@ -46,6 +50,7 @@ export function registerAllWidgets(PLASMIC: NextJsPlasmicComponentLoader) {
   PLASMIC.registerComponent(DynamicConnectorForm, DynamicConnectorFormMeta);
   PLASMIC.registerComponent(FeedbackWrapper, FeedbackWrapperMeta);
   PLASMIC.registerComponent(Markdown, MarkdownMeta);
+  PLASMIC.registerComponent(MarketingNavbar, MarketingNavbarMeta);
   PLASMIC.registerComponent(MonacoEditor, MonacoEditorMeta);
   PLASMIC.registerComponent(OSOChat, OSOChatMeta);
   PLASMIC.registerComponent(SupabaseWrite, SupabaseWriteMeta);
